chore(Explorers): moved API generators in specific folders
